+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/config/cache/AccCacheConfiguration.java
@@ -7,6 +7,8 @@ import org.springframework.context.annotation.Configuration;
 
 import eu.bcvsolutions.idm.acc.dto.AttributeValueWrapperDto;
 import eu.bcvsolutions.idm.acc.dto.SysProvisioningBreakItems;
+import eu.bcvsolutions.idm.acc.service.api.SynchronizationEntityExecutor;
+import eu.bcvsolutions.idm.acc.service.api.SynchronizationService;
 import eu.bcvsolutions.idm.acc.service.impl.AbstractSynchronizationExecutor;
 import eu.bcvsolutions.idm.acc.service.impl.DefaultSysProvisioningBreakConfigService;
 import eu.bcvsolutions.idm.core.api.config.cache.DistributedIdMCacheConfiguration;
@@ -28,7 +30,7 @@ public class AccCacheConfiguration {
 	 * @return IdMCacheConfiguration for {@link AbstractSynchronizationExecutor}
 	 */
 	@Bean
-	public IdMCacheConfiguration attributeMappingConfiguration() {
+	public IdMCacheConfiguration attributeMappingCacheConfiguration() {
 		return LocalIdMCacheConfiguration.<AttributeValueWrapperDto, Object> builder()
 				.withName(AbstractSynchronizationExecutor.CACHE_NAME)
 				.withKeyType(AttributeValueWrapperDto.class)
@@ -48,6 +50,20 @@ public class AccCacheConfiguration {
 				.withName(DefaultSysProvisioningBreakConfigService.CACHE_NAME)
 				.withKeyType(UUID.class)
 				.withValueType(SysProvisioningBreakItems.class)
+				.build();
+	}
+	
+	/**
+	 * Initialized synchronization executors.
+	 * 
+	 * @return
+	 */
+	@Bean
+	public IdMCacheConfiguration syncExecutorCacheConfiguration() {
+		return LocalIdMCacheConfiguration.<UUID, SynchronizationEntityExecutor> builder()
+				.withName(SynchronizationService.SYNC_EXECUTOR_CACHE_NAME)
+				.withKeyType(UUID.class)
+				.withValueType(SynchronizationEntityExecutor.class)
 				.build();
 	}